Uploading equilibrium index script
This script is a series of functions that takes in a list (or array) of numbers. It spits out the first equilibrium index that it finds. In other words, the sum of the numbers on either side of the index are equal in value. For an index at the beginning or end of the list, the sum is considered 0.